+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 + Added thicker window icons
 + Edits to "Glide" option in Long Jump
 + Added options "FPS", "TPS", "Potions" to Interface module
++ Use "Float.isNaN" check on yaw and pitch spoof rotations
 ```
 
 ---